backend/requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.40.16
botocore==1.40.16
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.2
openai==1.99.6
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
psycopg[binary]==3.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
twilio==9.7.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
exponent-server-sdk==2.2.0
gunicorn==21.2.0